refactor: move OwncastPlayer component
