+++ web/components/video/OwncastPlayer/OwncastPlayer.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { useHotkeys } from 'react-hotkeys-hook';
-import VideoJS from './player';
-import ViewerPing from './viewer-ping';
-import VideoPoster from './VideoPoster';
-import { getLocalStorage, setLocalStorage } from '../../utils/localStorage';
-import { isVideoPlayingAtom, clockSkewAtom } from '../stores/ClientConfigStore';
-import PlaybackMetrics from './metrics/playback';
-import createVideoSettingsMenuButton from './settings-menu';
-import LatencyCompensator from './latencyCompensator';
+import VideoJS from '../player';
+import ViewerPing from '../viewer-ping';
+import VideoPoster from '../VideoPoster';
+import { getLocalStorage, setLocalStorage } from '../../../utils/localStorage';
+import { isVideoPlayingAtom, clockSkewAtom } from '../../stores/ClientConfigStore';
+import PlaybackMetrics from '../metrics/playback';
+import createVideoSettingsMenuButton from '../settings-menu';
+import LatencyCompensator from '../latencyCompensator';
 
 const VIDEO_CONFIG_URL = '/api/video/variants';
 const PLAYER_VOLUME = 'owncast_volume';
