+++ web/pages/embed/video/index.tsx
@@ -8,7 +8,7 @@ import {
 } from '../../../components/stores/ClientConfigStore';
 import OfflineBanner from '../../../components/ui/OfflineBanner/OfflineBanner';
 import Statusbar from '../../../components/ui/Statusbar/Statusbar';
-import OwncastPlayer from '../../../components/video/OwncastPlayer';
+import OwncastPlayer from '../../../components/video/OwncastPlayer/OwncastPlayer';
 import { ClientConfig } from '../../../interfaces/client-config.model';
 import { ServerStatus } from '../../../interfaces/server-status.model';
